requirements.txt
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aioice==0.10.1
aiortc==1.13.0
aiosignal==1.3.2
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
arrow==1.3.0
asgiref==3.11.0
async-generator==1.10
attrs==23.1.0
av==14.4.0
backoff==2.2.1
beautifulsoup4==4.13.3
billiard==4.2.4
black==25.12.0
blinker==1.9.0
browser-use==0.1.40
cachetools==5.5.2
celery==5.6.0
certifi==2023.5.7
cffi==1.17.1
charset-normalizer==3.1.0
chromedriver-binary==114.0.5735.90.0
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.5
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
Django==5.2.9
django-cors-headers==4.9.0
django-tenants==3.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
dotenv==0.9.9
exceptiongroup==1.3.1
fastapi==0.110.2
filelock==3.18.0
flake8==7.3.0
Flask==3.1.0
flexcache==0.3
flexparser==0.4
frozenlist==1.6.0
fsspec==2025.5.0
future==1.0.0
gevent==24.11.1
gitdb==4.0.12
GitPython==3.1.45
google-crc32c==1.7.1
gotrue==2.12.0
greenlet==3.1.1
grpcio==1.72.1
grpcio-status==1.62.3
h11==0.14.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.34.3
hyperframe==6.1.0
idna==3.4
ifaddr==0.2.0
importlib_metadata==8.7.0
iniconfig==2.1.0
isort==7.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.6.1
langchain-anthropic==0.3.3
langchain-core==0.3.47
langchain-ollama==0.2.2
langchain-openai==0.3.1
langsmith==0.3.18
librt==0.7.5
llvmlite==0.44.0
loguru==0.7.3
lxml==6.0.2
Markdown==3.8
markdown2==2.5.4
markdownify==0.14.1
MarkupSafe==3.0.2
mccabe==0.7.0
monotonic==1.6
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.4.3
mypy==1.19.1
mypy_extensions==1.1.0
narwhals==2.0.1
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
ollama==0.4.7
openai==1.70.0
openai-whisper @ git+https://github.com/openai/whisper.git@dd985ac4b90cafeef8712f2998d62c59c3e62d22
orjson==3.10.15
outcome==1.2.0
packaging==24.2
pandas==2.0.2
pathspec==0.12.1
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==11.1.0
Pint==0.24.4
pipecat==0.3.0
pipecat-ai==0.0.69
platformdirs==4.3.8
playwright==1.51.0
plivo==4.59.3
pluggy==1.6.0
postgrest==1.0.1
posthog==3.21.0
prompt_toolkit==3.0.52
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.3
psutil==7.0.0
psycopg2-binary==2.9.11
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pyee==13.0.0
pyflakes==3.4.0
PyJWT==2.10.1
pylibsrtp==0.12.0
pyloudnorm==0.1.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
pypdfium2==5.1.0
PySocks==1.7.1
pytest==8.3.5
pytest-django==4.11.1
pytest-mock==3.14.0
python-dateutil==2.8.2
python-docx==1.2.0
python-dotenv==1.0.1
pytokens==0.3.0
pytz==2023.3
PyYAML==6.0.2
realtime==2.4.3
redis==7.1.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
resampy==0.4.3
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
scipy==1.15.3
selenium==4.10.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.6
soxr==0.5.0.post1
sqlparse==0.5.5
starlette==0.37.2
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.1
supafunc==0.9.4
sympy==1.14.0
tenacity==9.0.0
tiktoken==0.9.0
tokenizers==0.21.4
toml==0.10.2
torchaudio==2.7.1
tornado==6.5.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.3
twilio==9.0.1
types-python-dateutil==2.9.0.20250516
types-requests==2.32.4.20250913
typing_extensions==4.13.2
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.0.3
uv==0.6.9
uvicorn==0.29.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.14
webdriver-manager==3.8.6
websocket==0.2.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.0
zipp==3.23.0
zope.event==5.0
zope.interface==7.2
zstandard==0.23.0
